move path selector class to dedicated module.
